Refactor/#266: [TransactionItem] recalculateDateInRange 메서드 구현
